dts: seagate: Add Legend abstract board
This commit adds a Seagate directory in the dts directory and the Legend
devicetree which describe a common set of peripherals connected to a
STM32F070CB MCU and used by multiple boards. The goal of this devicetree
is to avoid duplication of information among boards that have nearly all
peripherals in common.

Signed-off-by: Maxime Bittan <maxime.bittan@seagate.com>
Signed-off-by: Simon Guinot <simon.guinot@seagate.com>